Seperate biasedFitness computation and purge as soon as max_population_size is reached
